Fixes to build the +couple version of QE
These changes are currently untested. A change in the COUPLE/CMakeLists.txt was
needed to install libqecouple.h. This change has been made in a new
branch, but this branch has not got a proper version number. As a result
Spack insists on picking the "generic" build system instead of "cmake". But
the "generic" (i.e. autotools) build system does not work for what we need.
Even trying to force Spack to use "cmake" hasn't worked. So, I guess I have
to wait until my QE change has been merged until I can actually test it in
a Spack environment.